[TASK] drag+drop in PageModule
this patch remove exception treatment when using drag+drop in PageModule
drag+drop now uses same cmd for copy/move/past as clipboard
s. https://forge.typo3.org/projects/typo3cms-core/repository/1749/revisions/c1be5540b20421fdfa295a1323b663f3189a41d7
