+++ Tests/Unit/Hooks/Datahandler/DatamapBeforeStartHookTest.php
@@ -54,53 +54,4 @@ class DatamapBeforeStartHookTest extends UnitTestCase
         self::assertIsArray($modDatamap['tt_content'][3]);
         self::assertSame(1, $modDatamap['tt_content'][3]['tx_container_parent']);
     }
-    /**
-     * @test
-     */
-    public function extractContainerIdFromColPosInDatamapSetsContainerIdToSplittedColPosValue(): void
-    {
-        $database = $this->prophesize(Database::class);
-        $containerFactory = $this->prophesize(ContainerFactory::class);
-        $dataHandlerHook = $this->getAccessibleMock(
-            DatamapBeforeStartHook::class,
-            ['foo'],
-            ['containerFactory' => $containerFactory->reveal(), 'database' => $database->reveal()]
-        );
-        $datamap = [
-            'tt_content' => [
-                39 => [
-                    'colPos' => '2-34',
-                    'sys_language_uid' => 0,
-                ],
-            ],
-        ];
-        $datamap = $dataHandlerHook->_call('extractContainerIdFromColPosInDatamap', $datamap);
-        self::assertSame(34, $datamap['tt_content'][39]['colPos']);
-        self::assertSame(2, $datamap['tt_content'][39]['tx_container_parent']);
-    }
-
-    /**
-     * @test
-     */
-    public function extractContainerIdFromColPosInDatamapSetsContainerIdToZeroValue(): void
-    {
-        $database = $this->prophesize(Database::class);
-        $containerFactory = $this->prophesize(ContainerFactory::class);
-        $dataHandlerHook = $this->getAccessibleMock(
-            DatamapBeforeStartHook::class,
-            ['foo'],
-            ['containerFactory' => $containerFactory->reveal(), 'database' => $database->reveal()]
-        );
-        $datamap = [
-            'tt_content' => [
-                39 => [
-                    'colPos' => '0',
-                    'sys_language_uid' => 0,
-                ],
-            ],
-        ];
-        $datamap = $dataHandlerHook->_call('extractContainerIdFromColPosInDatamap', $datamap);
-        self::assertSame(0, $datamap['tt_content'][39]['colPos']);
-        self::assertSame(0, $datamap['tt_content'][39]['tx_container_parent']);
-    }
 }
